requirements.txt
# Hard Drive Sound Simulator - Dependencias
# Python 3.8+

# Core dependencies
psutil>=5.9.0
pillow>=10.0.0
pystray>=0.19.0

# Build dependencies (para compilar)
pyinstaller>=6.0.0
